pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sign-language-overlay"
version = "0.1.0"
description = "Real-time sign language overlay for subtitles, audio, and live input"
requires-python = ">=3.9"
license = {text = "Custom - See LICENSE"}
authors = [{name = "Abdul Azeez Aris"}]
readme = "README.md"

dependencies = [
    "pygame>=2.5.0",
    "pysrt>=1.1.2",
    "webvtt-py>=0.4.6",
    "PyYAML>=6.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "mss>=9.0.0",
]

[project.optional-dependencies]
whisper = ["openai-whisper>=20231117", "torch>=2.0.0", "torchaudio>=2.0.0"]
realtime = ["sounddevice>=0.4.6"]
video = ["opencv-python>=4.8.0"]
dev = ["pytest>=7.0", "ruff>=0.1.0", "mypy>=1.0"]
all = [
    "sign-language-overlay[whisper,realtime,video,dev]",
]

[project.scripts]
sign-language-overlay = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
